Always write temporary projects for non-Maven sources
When compiling a .java file that is contained in, say, a Downloads/
directory, we definitely do not want to compile all of the .java files
available in the complete Downloads/** directory structure.

Now, we could special-case .java files containing the source for a class
in the default package, while still trying to discover .java files
recursively when the class is in a package and the directory structure
agrees with the package name.

But that would just violate the Law of the Least Surprise and confuse
people. So let's just make a temporary project in *every* case where we
do not find a pom.xml file in the expected location.

Incidentally, this makes sure that we always include pom.xml files in
the generated .jar files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>