Let the bindings accept the current ScriptModule instance
We recently improved the Scripting engine to accomodate the needs of
the ImageJ 1.x macro engine with regards to @Parameter annotations.
The major change was that the current ScriptModule instance is made
available to the engine bindings.

However, the Java script engine is a slightly different beast and
needs to make sure that no unexpected bindings are set (because they
cannot be used by the 'script').

Let's play nice and just ignore the binding.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>