Allow running inside Maven unit tests
When Maven's surefire plugin runs unit tests, the URLClassLoader does
actually not list the URLs corresponding to the dependencies. Instead,
surefire tries to be clever and hides those dependencies inside a
single, empty .jar file's manifest's Class-Path: entry.

We are not dumb, so surefire loses: we get the URLs anyway. Hah! Take
this, surefire! *shakes-fist*

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>