Verify that bare .java 'scripts' find all dependencies
... even when run via surefire...

To that end, we piggy-back on the already existing test for bare .java
"scripts": as scijava-common is a dependency of scripting-java, let's
just use one of scijava-common's utility classes in the bare .java
"script".  That way, the test will only pass when scijava-common's class
is found on the compiler class path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>