*Always* use the context class loader
If we ever limit which class loader we use as parent or to discover the
available dependencies by using the class loader that happened to load
MiniMaven, we will be in for a *lot* of trouble in the future, e.g. with
ImageJ 1.x' PluginClassLoader.

So let's not do that. Let's use the current thread's context class loader,
like everywhere else we need a class loader.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>